initial final commit for now

+++ StockApi/Modals/Product.cs
@@ -9,9 +9,9 @@ namespace StockApi.Modals
         [Required]
         public string Name { get; set; } = string.Empty;
         public int UnitMl { get; set; } = 750;
-        public float TareWeightG { get; set; } = 0;
-        public float LossAllowance { get; set; } = 0;
-        public float SellingPrice { get; set; }
+        public double TareWeightG { get; set; } = 0;
+        public double LossAllowance { get; set; } = 0;
+        public double SellingPrice { get; set; }
 
 
     }
